tourist-safety-system/backend/requirements.txt
# Core Flask Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Database: MongoDB (via pymongo)

# API and HTTP
requests==2.31.0
urllib3==2.0.7

# Cryptography and Security
cryptography==41.0.7
bcrypt==4.0.1

# Date and Time
python-dateutil==2.8.2

# JSON and Data Processing
simplejson==3.19.2

# Environment Variables
python-dotenv==1.0.0

# CORS Support
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# Logging
colorlog==6.7.0

# Geospatial
geopy==2.4.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Production Server
gunicorn==21.2.0
waitress==2.1.2
pymongo==4.8.0